requirements.txt
torch==2.2.2
torchvision==0.17.2
numpy==1.26.4
opencv-python
matplotlib
pyyaml
tqdm
Pillow
pycocotools


